requirements.txt
python-telegram-bot
python-dotenv
pydantic>=2.11.0,<3.0.0
